A-MoreOOP: Use More OOP
Some classes and main program contain long blocks of code and logics.

Code could be refactored to extract out closely related code
as classes.

Let's extract the classes below
* Ui.java, which handles interactions with user
* Storage.java, which handles loading and saving tasks in local file
* Parser.java, which handles user command with logics
* TaskList.java, which contains operations to modify task list
