+++ src/main/java/CaCaException.java
@@ -1,5 +1,6 @@
 /**
  * This is an exception class for exceptions specific to CaCa.
+ *
  * @author Carrie Zheng Jiarui
  * @version CS2103T AY22/23 Semester 1, iP
  */
