+++ src/main/java/Storage.java
@@ -4,14 +4,13 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Scanner;
 
 import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 
 /**
  * This class represents the storage system to save and read tasks from the hard disk.
+ *
  * @author Carrie Zheng Jiarui
  * @version CS2103T AY22/23 Semester 1, iP
  */
@@ -20,22 +19,34 @@ public class Storage {
     /**
      * A string representing the directory path.
      */
-    private static final String DIR_PATH = "data/";
+    private final String dirPath;
 
     /**
      * A string representing the file path.
      */
-    private static final String FILE_PATH = "data/caca.txt";
+    private final String filePath;
+
+    /**
+     * Constructor for creating a storage object.
+     *
+     * @param dirPath Directory path of file being read.
+     * @param filePath Path of the file being read.
+     */
+    public Storage(String dirPath, String filePath) {
+        this.dirPath = dirPath;
+        this.filePath = filePath;
+    }
 
     /**
      * Either displays all tasks stored in file (when applicable),
      * or prompts user if file does not exist.
+     *
      * @throws FileNotFoundException If file is not found.
      */
-    public static void loadFile() throws FileNotFoundException {
+    public void loadFile() throws FileNotFoundException {
         String LINE = "____________________________________________________________\n";
 
-        File file = new File(FILE_PATH);
+        File file = new File(filePath);
 
         if (file.exists()) {
 
@@ -67,12 +78,13 @@ public class Storage {
 
     /**
      * Creates a file if file does not exist in hard disk.
+     *
      * @param dirPath Directory path of created file.
      * @param filePath File path of created file.
      * @return The file created.
      * @throws IOException If there exists failed or interrupted I/O operations.
      */
-    public static File createFile(String dirPath, String filePath) throws IOException {
+    public File createFile(String dirPath, String filePath) throws IOException {
         File directory = new File(dirPath);
         File file = new File(filePath);
         directory.mkdir();
@@ -82,30 +94,35 @@ public class Storage {
 
     /**
      * Updates tasks in file.
+     *
      * @param tasks List of tasks in file.
      * @throws IOException If there exists failed or interrupted I/O operations.
      */
-    public static void updateFile(List<Task> tasks) throws IOException {
-        String newFilePath = FILE_PATH + ".new";
-        File file = createFile(DIR_PATH, newFilePath);
+    public void updateFile(TaskList tasks) throws IOException {
+        // Solution below adapted from W3.4c C++ to Java -> Miscellaneous Topics -> File access
+        // from https://nus-cs2103-ay2223s1.github.io/website/schedule/week3/topics.html#W3-4
+        // and https://github.com/cheehongw/ip/blob/master/src/main/java/duke/Storage.java
+        String newFilePath = filePath + ".new";
+        File file = createFile(dirPath, newFilePath);
         FileWriter fw = new FileWriter(file, true);
-        for (Task task : tasks) {
+        for (Task task : tasks.getTasks()) {
             fw.write(task.toFileFormat() + "\n");
         }
         fw.close();
 
         // Solution below adapted from https://docs.oracle.com/javase/tutorial/essential/io/move.html and
         // https://stackoverflow.com/questions/27749755/files-move-replace-existing-cannot-be-resolved-to-a-variable
-        Files.move(Paths.get(newFilePath), Paths.get(FILE_PATH), REPLACE_EXISTING);
+        Files.move(Paths.get(newFilePath), Paths.get(filePath), REPLACE_EXISTING);
     }
 
     /**
      * Reads tasks stored in file.
+     *
      * @return A list of all the tasks stored.
      */
-    public static List<Task> readFile() throws InvalidDateException {
-        List<Task> tasks = new ArrayList<>();
-        File file = new File(FILE_PATH);
+    public TaskList readFile() throws InvalidDateException {
+        TaskList tasks = new TaskList();
+        File file = new File(filePath);
 
         // Surround with try-catch block as suggested by IDE.
         Scanner sc = null;
@@ -115,8 +132,9 @@ public class Storage {
             while (sc.hasNextLine()) {
                 String task = sc.nextLine();
 
-                // Adapted from
+                // Solution below adapted from
                 // https://stackoverflow.com/questions/10796160/splitting-a-java-string-by-the-pipe-symbol-using-split
+                // and https://github.com/cheehongw/ip/blob/master/src/main/java/duke/Storage.java
                 String[] taskDetails = task.split(" \\| ");
                 String taskType = taskDetails[0];
                 String statusIcon = taskDetails[1];
@@ -126,15 +144,15 @@ public class Storage {
                 // Stores all the tasks from file into the array.
                 switch (taskType) {
                 case "T":
-                    tasks.add(new Todo(description, isDone));
+                    tasks.addTask(new Todo(description, isDone));
                     break;
                 case "D":
                     String by = taskDetails[3];
-                    tasks.add(new Deadline(description, by, isDone));
+                    tasks.addTask(new Deadline(description, by, isDone));
                     break;
                 case "E":
                     String at = taskDetails[3];
-                    tasks.add(new Event(description, at, isDone));
+                    tasks.addTask(new Event(description, at, isDone));
                     break;
                 default:
                     break;
@@ -145,7 +163,7 @@ public class Storage {
         } catch (FileNotFoundException e) {
             // Creates a new list to store subsequent tasks entered by user
             // if data file does not exist at the start.
-            tasks = new ArrayList<>();
+            tasks = new TaskList();
 
         } finally {
             // Prevents NullPointerException as suggested by IDE.
